easylist-backend/requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.0.0
gunicorn>=20.0.0
Pillow>=8.0.0
psycopg2>=2.9.0
